Make LevelBarProvidersPassive
When nodes are set to passive they stop preventing connected nodes from becoming inactive.
for example this is required for bluetooth speakers to allow switching away from passive outputs.

Also set name of nodes that will be filtered by program to "PulseAudio Volume Control",
since in 896a2793 node.name was changed to that for compatibility purposes.